Rewrite
Updated the loading of config options to be more.. dynamic I guess. This is mainly to reduce the amount of code but also to simplify things if more levels were added.
Renamed the gold option to money and changed it so it now reads copper. This will allow for more specific values. Maybe you want to send only a few silver to the player? Now you can!
Added the ability to send multiple items to the player. An explanation as to how that works is in the config file.